+++ instrumentation/ktor/ktor-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/ktor/v2_0/ServerInstrumentation.java
@@ -12,8 +12,8 @@ import io.ktor.server.application.Application;
 import io.ktor.server.application.ApplicationPluginKt;
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.instrumentation.ktor.internal.KtorBuilderUtil;
-import io.opentelemetry.instrumentation.ktor.server.AbstractKtorServerTracing;
-import io.opentelemetry.instrumentation.ktor.v2_0.server.KtorServerTracing;
+import io.opentelemetry.instrumentation.ktor.server.AbstractKtorServerTracingBuilder;
+import io.opentelemetry.instrumentation.ktor.v2_0.server.KtorServerTracingBuilderKt;
 import io.opentelemetry.javaagent.bootstrap.internal.AgentCommonConfig;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
@@ -40,19 +40,18 @@ public class ServerInstrumentation implements TypeInstrumentation {
 
     @Advice.OnMethodExit
     public static void onExit(@Advice.FieldValue("_applicationInstance") Application application) {
-      ApplicationPluginKt.install(application, KtorServerTracing.Feature, new SetupFunction());
+      ApplicationPluginKt.install(
+          application, KtorServerTracingBuilderKt.getKtorServerTracing(), new SetupFunction());
     }
   }
 
   public static class SetupFunction
-      implements Function1<AbstractKtorServerTracing.Configuration, kotlin.Unit> {
+      implements Function1<AbstractKtorServerTracingBuilder, kotlin.Unit> {
 
     @Override
-    public Unit invoke(AbstractKtorServerTracing.Configuration configuration) {
-      configuration.setOpenTelemetry(GlobalOpenTelemetry.get());
-      KtorBuilderUtil.serverBuilderExtractor
-          .invoke(configuration)
-          .configure(AgentCommonConfig.get());
+    public Unit invoke(AbstractKtorServerTracingBuilder builder) {
+      builder.setOpenTelemetry(GlobalOpenTelemetry.get());
+      KtorBuilderUtil.serverBuilderExtractor.invoke(builder).configure(AgentCommonConfig.get());
       return kotlin.Unit.INSTANCE;
     }
   }
