--- instrumentation/ktor/ktor-3.0/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v3_0/server/KtorServerTracing.kt
@@ -1,90 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.ktor.v3_0.server
-
-import io.ktor.server.application.*
-import io.ktor.server.request.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
-import io.ktor.util.*
-import io.ktor.util.pipeline.*
-import io.opentelemetry.context.Context
-import io.opentelemetry.extension.kotlin.asContextElement
-import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter
-import io.opentelemetry.instrumentation.api.semconv.http.HttpServerRoute
-import io.opentelemetry.instrumentation.api.semconv.http.HttpServerRouteSource
-import io.opentelemetry.instrumentation.ktor.internal.KtorServerTracingUtil
-import io.opentelemetry.instrumentation.ktor.server.AbstractKtorServerTracing
-import io.opentelemetry.instrumentation.ktor.v3_0.InstrumentationProperties.INSTRUMENTATION_NAME
-import kotlinx.coroutines.withContext
-
-class KtorServerTracing private constructor(
-  instrumenter: Instrumenter<ApplicationRequest, ApplicationResponse>,
-) : AbstractKtorServerTracing(instrumenter) {
-
-  companion object Feature : BaseApplicationPlugin<Application, Configuration, KtorServerTracing> {
-
-    private val contextKey = AttributeKey<Context>("OpenTelemetry")
-    private val errorKey = AttributeKey<Throwable>("OpenTelemetryException")
-
-    override val key: AttributeKey<KtorServerTracing> = AttributeKey("OpenTelemetry")
-
-    override fun install(pipeline: Application, configure: Configuration.() -> Unit): KtorServerTracing {
-      val configuration = Configuration(INSTRUMENTATION_NAME).apply(configure)
-
-      require(configuration.isOpenTelemetryInitialized()) { "OpenTelemetry must be set" }
-
-      val instrumenter = KtorServerTracingUtil.instrumenter(configuration)
-      val feature = KtorServerTracing(instrumenter)
-
-      val startPhase = PipelinePhase("OpenTelemetry")
-      pipeline.insertPhaseBefore(ApplicationCallPipeline.Monitoring, startPhase)
-      pipeline.intercept(startPhase) {
-        val context = feature.start(call)
-
-        if (context != null) {
-          call.attributes.put(contextKey, context)
-          withContext(context.asContextElement()) {
-            try {
-              proceed()
-            } catch (err: Throwable) {
-              // Stash error for reporting later since need ktor to finish setting up the response
-              call.attributes.put(errorKey, err)
-              throw err
-            }
-          }
-        } else {
-          proceed()
-        }
-      }
-
-      val postSendPhase = PipelinePhase("OpenTelemetryPostSend")
-      pipeline.sendPipeline.insertPhaseAfter(ApplicationSendPipeline.After, postSendPhase)
-      pipeline.sendPipeline.intercept(postSendPhase) {
-        val context = call.attributes.getOrNull(contextKey)
-        if (context != null) {
-          var error: Throwable? = call.attributes.getOrNull(errorKey)
-          try {
-            proceed()
-          } catch (t: Throwable) {
-            error = t
-            throw t
-          } finally {
-            feature.end(context, call, error)
-          }
-        } else {
-          proceed()
-        }
-      }
-
-      pipeline.monitor.subscribe(RoutingRoot.RoutingCallStarted) { call ->
-        HttpServerRoute.update(Context.current(), HttpServerRouteSource.SERVER, { _, arg -> arg.route.parent.toString() }, call)
-      }
-
-      return feature
-    }
-  }
-}
